move sketch file manipulation to own file
