+++ compiler/sketch-library/lib/index.js
@@ -1,77 +1,15 @@
 const fs = require('fs')
 const os = require('os')
 const path = require('path')
-const { Buffer } = require('buffer')
-const { exec, execSync } = require('child_process')
-const {
-  createNewSketchFile,
-  writeSketchFile,
-  generateId,
-} = require('sketch-file')
+const { exec } = require('child_process')
 const renderDocument = require('./render-document')
+const modifySketchTemplate = require('./modify-sketch-template')
 
 // Temporary directory for the compiler to write generated JS files
 const compilerOutput = path.join(os.tmpdir(), 'lona-sketch-library-generated')
 
-function findImages(layers) {
-  let images = {}
-  layers.forEach(layer => {
-    if (layer && layer.style && layer.style.fills) {
-      layer.style.fills.forEach(fill => {
-        if (!fill.image) {
-          return
-        }
-        if (fill.image.data && fill.image.sha1) {
-          images[fill.image.sha1._data] = Buffer.from(
-            fill.image.data._data,
-            'base64'
-          )
-          fill.image._ref = 'images/' + fill.image.sha1._data
-          delete fill.image.data
-          delete fill.image.sha1
-          fill.image._class = 'MSJSONFileReference'
-        }
-      })
-    }
-    if (layer.layers) {
-      Object.assign(images, findImages(layer.layers))
-    }
-  })
-  return images
-}
-
-function modifySketchTemplate({ layers, textStyles, colors }, output) {
-  const sketchDoc = createNewSketchFile(generateId(output))
-
-  const images = findImages(layers)
-
-  sketchDoc.document.layerTextStyles.objects = textStyles
-  sketchDoc.document.assets.colors = colors.map(c => ({
-    _class: 'color',
-    alpha: c.alpha,
-    blue: c.blue,
-    green: c.green,
-    red: c.red,
-  }))
-  sketchDoc.document.assets.colorAssets = colors.map(c => ({
-    _class: 'MSImmutableColorAsset',
-    name: c.name,
-    color: {
-      _class: 'color',
-      alpha: c.alpha,
-      blue: c.blue,
-      green: c.green,
-      red: c.red,
-    },
-  }))
-  sketchDoc.pages[0].layers = sketchDoc.pages[0].layers.concat(layers)
-  sketchDoc.images = images
-
-  return writeSketchFile(sketchDoc, output)
-}
-
 module.exports = function(workspace, sketchFilePath, options) {
-  const { devicePresetList, compiler, componentPathFilter, logFunction } =
+  let { devicePresetList, compiler, componentPathFilter, logFunction } =
     options || {}
 
   function log(text) {
